Moved some utility classes to a separate file
